TTS/Indic-F5/requirements.txt
soundfile
librosa
transformers==4.49.0
torchcodec==0.8.1